pyproject.toml
[dependency-groups]
dev = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.2",
    "bm25s>=0.3.0",
    "chonkie>=1.5.6",
    "chromadb>=1.5.1",
    "dspy>=3.1.3",
    "fire>=0.7.1",
    "flake8>=7.3.0",
    "langchain>=1.2.10",
    "langchain-text-splitters>=1.1.1",
    "mypy>=1.19.1",
    "onnxruntime<1.24",
    "pydantic>=2.12.5",
    "pystemmer>=3.0.0",
    "python-fire>=0.1.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
]
